requirements.txt
Flask==3.0.*
flask-apscheduler==1.13.*
flask-cors==4.0.*
flask-limiter==3.5.*
azure-identity==1.15.*
azure-cosmos==4.5.*
PyJWT[crypto]==2.8.*
google-generativeai==0.3.*
pydantic==2.5.*
python-dotenv==1.0.*
sendgrid==6.11.*
jinja2==3.1.*
requests==2.31.*
redis==5.0.*
gunicorn==21.2.*
pytest==7.4.*
pytest-mock==3.12.*
vcrpy==5.1.*
locust==2.17.*
